Refactor EmptyState & ExpandableInlineSearch incubator components.
