+++ src/main/java/gwt/material/design/incubator/client/search/InlineSearchClientBundle.java
@@ -17,15 +17,15 @@
  * limitations under the License.
  * #L%
  */
-package gwt.material.design.incubator.client.emptystate;
+package gwt.material.design.incubator.client.search;
 
 import com.google.gwt.core.client.GWT;
 import com.google.gwt.resources.client.ClientBundle;
 import com.google.gwt.resources.client.TextResource;
 
-public interface EmptyStateDebugClientBundle extends ClientBundle {
-    EmptyStateDebugClientBundle INSTANCE = GWT.create(EmptyStateDebugClientBundle.class);
+public interface InlineSearchClientBundle extends ClientBundle {
+    InlineSearchClientBundle INSTANCE = GWT.create(InlineSearchClientBundle.class);
 
-    @Source("resources/css/empty-state.css")
-    TextResource emptyStateDebugCss();
+    @Source("resources/css/inline-search.min.css")
+    TextResource inlineSearchCss();
 }
