--- src/main/resources/gwt/material/design/incubator/client/emptystate/resources/css/empty-state.css
@@ -1,28 +1,0 @@
-.empty-state {
-    text-align: center;
-    padding-top: 60px;
-}
-
-.empty-state :nth-child(1) {
-    width: 100px;
-    border-radius: 50%;
-    padding: 20px;
-    height: 100px;
-    background: #bbdefb;
-    margin: auto;
-}
-
-.empty-state :nth-child(2) {
-    font-size: 2em !important;
-    text-align: center !important;
-    margin-top: 20px !important;
-    color: dimgrey !important;
-}
-
-.empty-state :nth-child(3) {
-    font-size: 0.8em !important;
-    width: 80% !important;
-    text-align: center !important;
-    color: #9a9a9a !important;
-    margin: auto !important;
-}
