+++ src/main/resources/gwt/material/design/incubator/client/search/resources/css/inline-search.css
